Crate bdk_core

This crate is a collection of core structures for [Bitcoin Dev Kit].

Re-exports

pub extern crate serde;

pub use bitcoin;

Modules

spk_client

Helper types for spk-based blockchain clients.

Structs

BlockId

A reference to a block in the canonical chain.

CheckPoint

A checkpoint is a node of a reference-counted linked list of BlockIds.

CheckPointIter

Iterates over checkpoints backwards.

ConfirmationBlockTime

Represents the confirmation block and time of a transaction.

TxUpdate

Data object used to communicate updates about relevant transactions from some chain data source to the core model (usually a bdk_chain::TxGraph).

Traits

Merge

Trait that makes an object mergeable.

Type Aliases

Indexed

A tuple of keychain index and T representing the indexed value.

KeychainIndexed

A tuple of keychain K, derivation index (u32) and a T associated with them.