Struct bdk_core::TxUpdate

pub struct TxUpdate<A = ()> {
    pub txs: Vec<Arc<Transaction>>,
    pub txouts: BTreeMap<OutPoint, TxOut>,
    pub anchors: BTreeSet<(A, Txid)>,
    pub seen_ats: HashMap<Txid, u64>,
}

Data object used to communicate updates about relevant transactions from some chain data source to the core model (usually a bdk_chain::TxGraph).

Fields
txs: Vec<Arc<Transaction>>

Full transactions. These are transactions that were determined to be relevant to the wallet given the request.

txouts: BTreeMap<OutPoint, TxOut>

Floating txouts. These are TxOuts that exist but the whole transaction wasn’t included in txs since only knowing about the output is important. These are often used to help determine the fee of a wallet transaction.

anchors: BTreeSet<(A, Txid)>

Transaction anchors. Anchors tells us a position in the chain where a transaction was confirmed.

seen_ats: HashMap<Txid, u64>

Seen at times for transactions. This records when a transaction was most recently seen in the user’s mempool for the sake of tie-breaking other conflicting transactions.

Implementations

impl<A: Ord> TxUpdate<A>

pub fn map_anchors<A2: Ord, F: FnMut(A) -> A2>(self, map: F) -> TxUpdate<A2>

Transforms the TxUpdate to have anchors (A) of another type (A2).

This takes in a closure with signature FnMut(A) -> A2 which is called for each anchor to transform it.

pub fn extend(&mut self, other: TxUpdate<A>)

Extend this update with other.